+++ src/tool/indexing.ml
@@ -658,8 +658,7 @@ module UserLevelQueries = struct
    let resultsno = List.length items in
    let _,items = Lplib.List.cut items from in
    let items,_ = Lplib.List.cut items how_many in
-   Format.asprintf
-    "<h1>Number of results: %d</h1>%a@."
+   Format.asprintf "<h1>Number of results: %d</h1>%a@."
     resultsno pp_results items
   with
    | Stream.Failure ->